requirements.txt
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.22.0
requests==2.32.3
